Improve test parallelism by moving some slow tests
Moving the slow tests to a new class will allow more tests to be
run in parallel because testing parallelism in Apache maven is at
the class level.
